EntityUIBaseにbuffのアニメーションを追加
テストしてないので、どんな感じかわ知らん

+++ TOC_Game/Assets/Code/C#/Entity/Player/PlayerUISystem_Gabu.cs
@@ -26,7 +26,7 @@ public class PlayerUISystem_Gabu : EntityUIBase
     }
 
 
-    public override void Buffed(long value)
+    public override void Buffed(float value)
     {
 
     }
